Revert "feat: use u64 which are supposedly always loaded atomcally"
This reverts commit ae9626141ad41083055d5f61e268a9efd63efac2.

+++ memcached-ebpf-proxy-cache-common/src/lib.rs
@@ -58,7 +58,7 @@ impl Hasher for Fnv1AHasher {
 
 #[repr(C)]
 pub struct CacheEntry {
-    pub token: u64,
+    pub token: u32,
     pub len: u32,
     pub valid: bool,
     pub hash: u32,
